+++ src/main/java/com/spring/practice/rest/examples/controller/ResponseStatusController.java
@@ -1,4 +1,4 @@
-package com.spring.practice.rest.controller.exceptions;
+package com.spring.practice.rest.examples.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -25,7 +25,7 @@ public class ResponseStatusController {
         try {
             throw new Exception("Bad Request");
         } catch (Exception e) {
-            log.error(String.format("status code=%d", HttpStatus.BAD_REQUEST));
+            log.error(String.format("status code=%d", HttpStatus.BAD_REQUEST.value()));
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "bad request error");
         }
     }
